requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
pyyaml>=6.0
pymysql>=1.1.0
requests>=2.31.0
